pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airstudio"
version = "0.1.0"
description = "Local Gradio UI for AirLLM with per-request metrics"
readme = "airstudio.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "AirStudio" }]

dependencies = [
  "airllm>=2.11.0",
  "gradio>=4.0.0",
  "nvidia-ml-py>=12.535.108",
  "optimum>=1.21.0,<2.0.0",
  "psutil>=5.9.0",
  "pyyaml>=6.0",
  "sentencepiece>=0.2.0",
  "torch",
  "transformers>=4.37.0,<4.49.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
